requirements/base.txt
asyncio-redis==0.16.0
boto3==1.14.56
faust==1.10.4
python-schema-registry-client==1.4.6
